update trimerge array, object, and map to handle internal equality
With this change, when trimerging an object, array, or map, if the final
merged object/array/map has the same keys and values as left or right,
one of those two will be returned.

This will obviate the need to have a deep equality merger in every merge
chain, though you still want the simple `trimergeEquality` merger.
