+++ src/index.ts
@@ -9,11 +9,7 @@ export {
   CannotMergeError,
   MergeFn,
 } from './trimerge';
-export {
-  trimergeArrayCreator,
-  trimergeJsonDeepEqual,
-  trimergeJsonObject,
-} from './trimerge-json';
+export { trimergeJsObject } from './trimerge-js-object';
 export {
   routeMergers,
   RouteWildCard,
@@ -39,3 +35,5 @@ export {
 } from './node-diff3';
 export { diff3MergeStringRanges } from './diff3-string';
 export { diff3Keys } from './diff3-keys';
+export { trimergeArrayCreator } from './trimerge-js-array';
+export { trimergeJsonDeepEqual } from './trimerge-json-equal';
