+++ src/trimerge-js-array.test.ts
@@ -1,190 +1,8 @@
-import {
-  CannotMergeError,
-  combineMergers,
-  MergeFn,
-  trimergeEquality,
-} from './trimerge';
-import { Path } from './path';
-import { CannotMerge } from './cannot-merge';
-import {
-  trimergeArrayCreator,
-  trimergeJsonDeepEqual,
-  trimergeJsonObject,
-} from './trimerge-json';
+import { CannotMergeError, combineMergers, trimergeEquality } from './trimerge';
+import { trimergeArrayCreator } from './trimerge-js-array';
+import { trimergeJsonDeepEqual } from './trimerge-json-equal';
 
-describe('trimergeJsonDeepEqual', () => {
-  it('merges left === right', () => {
-    const s1 = { state: 1 };
-    const s2 = { state: 3 };
-    const s3 = { state: 3 };
-    const initialmerger = jest.fn(() => CannotMerge);
-    const merger = combineMergers(initialmerger, trimergeJsonDeepEqual);
-    expect(merger(s1, s2, s3)).toBe(s2); // defaults to left
-    expect(initialmerger.mock.calls).toEqual([[s1, s2, s3, [], merger]]);
-  });
-  it('merges orig === left', () => {
-    const s1 = { state: 1 };
-    const s2 = { state: 1 };
-    const s3 = { state: 3 };
-    const merger = combineMergers(trimergeJsonDeepEqual);
-    expect(merger(s1, s2, s3)).toBe(s3);
-  });
-  it('merges orig === right', () => {
-    const s1 = { state: 1 };
-    const s2 = { state: 2 };
-    const s3 = { state: 1 };
-    const merger = combineMergers(trimergeJsonDeepEqual);
-    expect(merger(s1, s2, s3)).toBe(s2);
-  });
-});
-
-function mockPathTrackingMerger(paths: Path[]): MergeFn {
-  return (_orig, _left, _right, path = []): typeof CannotMerge => {
-    paths.push(path);
-    return CannotMerge;
-  };
-}
-
-describe('trimergeJsonObject', () => {
-  it('adds field', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1, world: 2, there: 3 };
-    const s3 = { hello: 1, world: 2 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({ hello: 1, world: 2, there: 3 });
-    expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
-  });
-  it('adds two fields', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1, world: 2, there: 3 };
-    const s3 = { hello: 1, world: 2, here: 4 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({
-      hello: 1,
-      world: 2,
-      there: 3,
-      here: 4,
-    });
-    expect(paths).toEqual([[], ['hello'], ['world'], ['there'], ['here']]);
-  });
-  it('adds against undefined', () => {
-    const s1 = undefined;
-    const s2 = { hello: 1, there: 3 };
-    const s3 = { hello: 1, here: 4 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({
-      hello: 1,
-      there: 3,
-      here: 4,
-    });
-    expect(paths).toEqual([[], ['hello'], ['there'], ['here']]);
-  });
-  it('changes field', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1, world: 2 };
-    const s3 = { hello: 1, world: 3 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({ hello: 1, world: 3 });
-    expect(paths).toEqual([[], ['hello'], ['world']]);
-  });
-  it('removes field', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1 };
-    const s3 = { hello: 1, world: 2 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({ hello: 1 });
-    expect(paths).toEqual([[], ['hello'], ['world']]);
-  });
-  it('adds and changes field', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1, world: 2, there: 2 };
-    const s3 = { hello: 1, world: 3 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({ hello: 1, world: 3, there: 2 });
-    expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
-  });
-  it('adds and removes field', () => {
-    const s1 = { hello: 1, world: 2 };
-    const s2 = { hello: 1, world: 2, there: 2 };
-    const s3 = { hello: 1 };
-    const paths: Path[] = [];
-    const merger = combineMergers(
-      mockPathTrackingMerger(paths),
-      trimergeEquality,
-      trimergeJsonObject,
-    );
-    expect(merger(s1, s2, s3)).toEqual({ hello: 1, there: 2 });
-    expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
-  });
-  it('does not merge if not all objects 1', () => {
-    const s1 = false;
-    const s2 = {};
-    const s3 = {};
-    const merger = combineMergers(trimergeJsonObject);
-    expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
-  });
-  it('does not merge if not all objects 2', () => {
-    const s1 = {};
-    const s2 = false;
-    const s3 = {};
-    const merger = combineMergers(trimergeJsonObject);
-    expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
-  });
-  it('does not merge if not all objects 3', () => {
-    const s1 = {};
-    const s2 = {};
-    const s3 = false;
-    const merger = combineMergers(trimergeJsonObject);
-    expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
-  });
-  it('does not merge if none are objects', () => {
-    const s1 = false;
-    const s2 = false;
-    const s3 = false;
-    const merger = combineMergers(trimergeJsonObject);
-    expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
-  });
-  it('does not class instances', () => {
-    class Foo {}
-    const s1 = new Foo();
-    const s2 = new Foo();
-    const s3 = new Foo();
-    const merger = combineMergers(trimergeJsonObject);
-    expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
-  });
-});
-
-describe('arrays', () => {
+describe('trimergeArrayCreator', () => {
   const basicArrayMerge = combineMergers(
     trimergeEquality,
     trimergeArrayCreator((item) => String(item)),
