+++ src/trimerge-json-equal.ts
@@ -1,0 +1,4 @@
+import { trimergeEqualityCreator } from './trimerge';
+import { jsonEqual } from './json-equal';
+
+export const trimergeJsonDeepEqual = trimergeEqualityCreator(jsonEqual);
